Add comments to lib/recoil, refactor state sync
